Se realizan cambios para ejecutar bot desde un thread para poner anidar varios
