+++ SharkTeethCastleBot/languages/strings_es.py
