+++ SharkTeethCastleBot/services/auth.py
