+++ SharkTeethCastleBot/utils/constants.py
