--- pyvenv.cfg
@@ -1,8 +1,0 @@
-home = /usr
-implementation = CPython
-version_info = 3.8.5.final.0
-virtualenv = 20.0.21
-include-system-site-packages = false
-base-prefix = /usr
-base-exec-prefix = /usr
-base-executable = /usr/bin/python
