+++ requirements.txt
@@ -6,7 +6,7 @@ idna==2.9
 markdown-strings==3.3.0
 pika==1.1.0
 pymongo==3.10.1
-pyTelegramBotAPI==3.7.1
+pyTelegramBotAPI==3.7.3
 pytz==2020.1
 requests==2.24.0
 schedule==0.6.0
